imagenet/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.0=mkl
blosc=1.20.1=hd408876_0
brotli=1.0.9=he6710b0_2
brotli-python=1.0.9=py37hb892b2f_3
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.7.22=hbcca054_0
cachecontrol=0.13.1=pyhd8ed1ab_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.14.4=py37h261ae71_0
charls=2.1.0=he6710b0_2
charset-normalizer=3.2.0=pyhd8ed1ab_0
click=7.1.2=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
colorama=0.4.6=pyhd8ed1ab_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.1.1=pyhd3eb1b0_0
decorator=4.4.2=pyhd3eb1b0_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
freetype=2.10.4=h5ab3b9f_0
giflib=5.1.4=h14c3975_1
hdmedians=0.13=py37h3010b51_1000
helper=2.5.0=pypi_0
idna=3.4=pyhd8ed1ab_0
imagecodecs=2021.1.11=py37h581e88b_1
imageio=2.9.0=py_0
importlib-metadata=4.11.4=py37h89c1867_0
importlib_metadata=4.11.4=hd8ed1ab_0
iniconfig=2.0.0=pyhd8ed1ab_0
intel-openmp=2020.2=254
ipython=7.33.0=py37h89c1867_0
jedi=0.19.0=pyhd8ed1ab_0
joblib=1.0.0=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jxrlib=1.1=h7b6447c_2
kiwisolver=1.3.1=py37h2531618_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
lerc=2.2.1=h2531618_0
libaec=1.0.4=he6710b0_1
libdeflate=1.7=h27cfd23_5
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
libwebp=1.0.1=h8e7db2f_0
libzopfli=1.0.3=he6710b0_0
lockfile=0.12.2=py_1
lz4-c=1.9.3=h2531618_0
matplotlib-base=3.3.2=py37h817c723_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
msgpack-python=1.0.0=py37hc928c03_2
natsort=7.1.1=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
networkx=2.5=py_0
ninja=1.10.2=py37hff7bd54_0
nltk=3.5=py_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=py37_0
opencv-python=4.5.1.48=pypi_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1h=h516909a_0
packaging=23.1=pyhd8ed1ab_0
pandas=1.2.3=py37ha9443f7_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=7.2.0=pypi_0
pip=20.3.3=py37h06a4308_0
pluggy=1.0.0=py37h89c1867_3
prompt-toolkit=3.0.39=pyha770c72_0
protobuf=3.12.2=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pycparser=2.20=py_2
pygments=2.16.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyhd3eb1b0_0
pysocks=1.7.1=py37h89c1867_5
pytest=7.4.0=pyhd8ed1ab_0
python=3.7.7=hcff3b4d_5
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.7=1_cp37m
pytorch=1.7.1=py3.7_cuda10.1.243_cudnn7.6.3_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.4.1=py37h27cfd23_1
readline=8.1=h27cfd23_0
regex=2020.11.13=py37h27cfd23_0
requests=2.31.0=pyhd8ed1ab_0
scikit-bio=0.5.6=py37ha21ca33_4
scikit-image=0.17.2=py37h0da4684_1
scikit-learn=0.24.2=py37ha9443f7_0
scipy=1.6.0=py37h91f5cce_0
seaborn=0.12.2=py37h06a4308_0
setuptools=52.0.0=py37h06a4308_0
six=1.15.0=py37h06a4308_0
snappy=1.1.8=he6710b0_0
sqlite=3.33.0=h62c20be_0
tensorboard-logger=0.1.0=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tifffile=2021.1.14=pyhd3eb1b0_1
tk=8.6.10=hbc83047_0
tomli=2.0.1=pyhd8ed1ab_0
toolz=0.11.1=pyhd3eb1b0_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cu101
tornado=6.1=py37h27cfd23_0
tqdm=4.56.0=pyhd3eb1b0_0
traitlets=5.9.0=pyhd8ed1ab_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=2.0.4=pyhd8ed1ab_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zfp=0.5.5=h2531618_4
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
